requirements.txt
# Core Dependencies
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Trading & Market Data
ccxt>=4.0.0
ta>=0.11.0

# Data Analysis
matplotlib>=3.8.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0

# Logging & Monitoring
colorlog>=6.7.0

# Optional: For full ML features (can install later)
# tensorflow>=2.15.0

